fix: migrated throttle back to waitForEvent, moved waitForEvent to shared utils
